backend_django/quizApplication/requirements.txt
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1610094662606/work
argon2-cffi @ file:///Users/runner/miniforge3/conda-bld/argon2-cffi_1625821429979/work
asgiref==3.4.1
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
autopep8==1.6.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.11.1
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1629908509068/work
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.2.0
certifi==2021.5.30
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1625835350093/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1610093454858/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1626371162869/work
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conda==4.10.3
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1618231726924/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1616851494586/work
cycler==0.11.0
Cython==0.29.26
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1627074942557/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1621187651333/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.8.1
django-mssql==1.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
fonttools==4.34.4
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1609836280497/work
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1630267518765/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1630405734294/work/dist/ipykernel-6.3.1-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1630172565945/work
ipython-genutils==0.2.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1630470352784/work
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1610146802570/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1621419064915/work
joblib==1.1.0
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1614815863336/work
jupyter @ file:///Users/runner/miniforge3/conda-bld/jupyter_1611871906259/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1630336909568/work
jupyter-console @ file:///home/conda/feedstock_root/build_artifacts/jupyter_console_1616560109969/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1612125274649/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1630474294342/work
kiwisolver==1.4.3
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1621455750847/work
matplotlib==3.5.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1618935594181/work
mistune @ file:///Users/runner/miniforge3/conda-bld/mistune_1624941595202/work
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1629120697898/work
nbconvert @ file:///Users/runner/miniforge3/conda-bld/nbconvert_1624472693960/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
nltk==3.7
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1628690102344/work
numpy==1.23.1
oauthlib==3.2.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1625323647219/work
pandas==1.4.3
pandocfilters==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.2.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1622586138406/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1629903925368/work
protobuf==4.21.2
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1624996792650/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
PyJWT==2.3.0
pymssql @ git+https://github.com/pymssql/pymssql@f486f7d79bd2583eba96323cdeca34eb8f14cd2a
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
pyrsistent @ file:///Users/runner/miniforge3/conda-bld/pyrsistent_1624984668772/work
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291460269/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1628233219441/work
regex==2022.7.9
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1626393743643/work
requests-oauthlib==1.3.1
rsa==4.8
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1624999868738/work
Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1628511208346/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.3.2.post1
SQLAlchemy==1.4.28
sqlparse==0.4.2
terminado @ file:///Users/runner/miniforge3/conda-bld/terminado_1629328258161/work
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1621261527237/work
toml==0.10.2
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1625488956802/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1628985621016/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1630423529112/work
uritemplate==4.1.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1624634538755/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
widgetsnbextension @ file:///Users/runner/miniforge3/conda-bld/widgetsnbextension_1605475516388/work
zipp==3.6.0
